Pulled in routing info, defined srm templates.
Need to get srm behavior in place; should be simple.
For now, we have no remissioning capability; there
are some complexities that need thinking through.
For SRM, the biggest thing is that when units
hit R anything, they're deployable/missionable.
If we do that, we get the ability to hook into
the existing marathon fill routines (deployable
supply and the like).  Added a new behavior
called :mission, since SRM units may or may not bog
or do other wierd stuff.  Their behavior is entirely
dependent on the demand definition.  The demand
calls the shots...so the unit behavior is responsible
for interpreting what happens.