revised create-units to handle multiple ents
Now we use a base-record straight from the data
when we make multiple units, and re-use
record->unitdata to handle the minutae.
Additionally, this fixes some disparities
that were showing up in the handling of
new SRM types, where we have units that
are SRM, but due to a batch record with
quantity >1, they end up not going through
the SRM creation pipeline.  They do now,
so yay.