cleaned up behavior messaging.
Messages are now deferred for handling
in an update cycle until the entity is
synchronized with the current time.
So, we perform a full roll-forward
behavior evaluation, now with an
update-state-beh that does not
process messages.  At the end of
rolling forward, we're synchronized
with the time of the pending message,
so we can handle it.  Basically we
check messages after rolling, as we
should have been doing.