Added initial port of observers to capture deltas.
Observers are currently under construction, and
serve primarily as a means to derive concrete
data from ephemeral events.  The legacy strategy
was to have observers log activities to their own
internal state, typically writing some serialized
form of event to output, or computing stats before
writing.

Now, observers are more like "active" systems in
the ECS connotation.  They are invoked upon processing
an event - functionally - and compute the resulting
context.  By convention, the current set of observers
watch deployments and entity movement, recording
the instantenous deltas until such as time as the
system has stabilized.  This occurs at the end of
the day, which causes an event to fire that notifies
them.  At the end of the day, the ephemeral deltas
are persisted to the state, which is consistent with
generating a persistent "history" from which we
can derive other histories.  At the beginning of
each day, any ephemeral components in the entity
store are dropped, so that we only capture
observered changes via the presence of related
components each day.  We can view the observers
as a form of middleware for the system.