Performance hack
Added a wrapper in marathon.ces.core/trigger-event.
A non-trivial bottleneck in some operations is
triggering events, particularly building strings
for event messages.
Since event messages only matter for humans, and
if we're actually interested in reading them,
we can defer the work need to build them.

So, core/trigger-event, which is a drop-in
replacement for spork.sim.simcontext/trigger-event
allows us to check our debugging status
by looking at core/*debug*.  If the binding is
true, we allow the msg form to persist through
to evaluation.  If not, we ignore it and
don't do the work (at least not the work of
building the string implied by the form).

We got about a 100ms improvement in speed out of
this.  Looking at further opportunities to reduce
capacity runs.  Remaining bottlenecks include
supply updates (amenable to parallelization).