Correcting behaviors related to policy / withdraw
Significant changes to debugging ala
marathon.analysis/entity-trace, to support
gleaning a visual history from the point-of-error.

Scrubbed out needless message-handling of
superfluous entity-update msg, and accounted for
an updated last-update key in up-to-date
behavior.  This saves on processing time and
trims down logged ouput.

Working to identify the source of seemingly
duplicated state changes due to an abrupt
withdraw.  Current suspects are
supply/release-followon-unit and
demand/withdraw-unit .