Small compliance patches plus big init fix
The bulk of the changes are to ensure
compliance with legacy formats.  However,
at least two major bugs have been
patched:

1) marathon.ces.engine/initialize-sim
We were supposed to be using this, and
were in a couple of tests, since it provides
all the hooks for scheduling things and
computing the last day, etc.
Unfortunately, it was left out of the
primary API for context creation in
marathon.analysis/load-context.

So, the major consequence was the lack
of guaranteeing scheduling policy updates
for the days in which periods change.

2) Bug in marathon.ces.behavior,
position=location?, due to faulty
implementation in function some-member.
We were returning a truthy value
regardless of input, rather than
testing set membership.  This caused
problems with overlapping positions
causing an apparent location change.
Now, overlapping keeps the unit in the
same location while changing the
policy position.