Fixed apparent error with demandtrends
Demandtrends files should be identical,
they werent'.  The demand profile in
m4 was off due to undersampling,
specifically onl sampling trends
on days when there were demand changes
as registered by quantity changes.
In particular, we need to capture the
demand's information immediately PRIOR
to deactivation, that is, we need the
'end points' in the discrete signal,
so endday - 1.  This gives proc something
to work with when it's converting
demandtrends into discrete samples.
Currently, we closer to sampling
- redundantly - as legacy version
does.  We only need to capture
discrete changes and let the
post processor sort it out,
since it's designed to reconstruct
the signal from sparse samples,
so long as we have all the inflection
points covered.  We do now, and the
demand profiles look identical.
We'll look at sparser demand trend
sampling as a useful optimization
in the future - i.e. only sample
demands that change, just activated,
or are about to deactivate, rather
than the current "sample all active
demands" approach, regardless of
specific changes.