Fixed faux unit count leading to incorrect overlap
The marathon.ces.demand/unit-count function only
accounted for :units-assigned, and not
:units-overlapping.  When we went to deactivate
a demand, and see if any units needed to go home,
our demand appeared empty because of this.  Thus,
overlapping units never received the "send home"
behavior to dissociate them from the demand -
although internally, they had already left the
demand location.  We could put in a runtime
error check to ensure this, or some specific
location-insurance regression tests to make
sure that, if a unit says it's at a location,
and the location is a demand, then that
demand has the unit registered as a member.