Fixed lack of complete followon fills
We had a problem fully exhausting the followon supply during the call
to `marathon.ces.fill.demand/fill-followons`, in which only the first
group of followon eligible demands was filled.  The behavior was more
akin to the "early stop" of the hierarchical fill behavior than the
"plow on through" behavior of followon-fill.  Error was a mistyped
keyword predicate indicating early-stop, `:early-stop?` which was sent
as `:early-stop`. Problem solved.