Inserted runtime checks into sorting predicates
nil is no longer an acceptable sorting criterion, at least not by
default.  To address and detect the missing followon code from before,
we now define marathon.ces.query/get! and use it in all sorting
functions, rather than get.  get! will throw an exception if nil is
returned from the inner call to get, ensuring that we are made aware
of missing sort keys.

After re-running with get!, and enabling the previous when-followon
sorting criterion, no errors occurred.  We assume that the fix to
wait-time in the previous patch fixed the underlying problem of empty
followoncode.