non-bog implemented, policy-change stable
Most of the infrastructure for non-bog effects are present.  The only
difference between current and legacy implementations are that the
legacy implementation includes any non-boggable supply in the fill
buckets.  Right now, policy constraints are keeping non-bog supply out
of circulation, but that's an easy fix.

Deployments are matching perfectly for BCT types for varying policies,
for reference quantities from base case, and across both atomic and
composite policies.  Discovered and fixed another numerical stability
error -- very minor -- that was creating sparse off-by-one errors when
projecting between policies during policy-change states.  Corrected
this by re-using the useful trunc-float function to ensure a truncated
decimal accuracy to within 6 decimal places.  The error was primarily
a VBA-ism, but this seems to have fixed it.

Loosening NonBOG availability constraints appears to be one of, if not
the, final obstacle for BCT-wide verification.