Minor changes for performance, prepping nonbog
1) Examination of marathon.ces.query yielded
some low-hanging performance optimization
fruit.  Specifically, if we know the set
of rules we're interested in a-priori,
there's no need to look at every rule and
filter.  Merely select the appropriate
rules during traversal of supply buckets.
For smallish simulations, with few srcs,
this won't matter too much, but for the
large legacy runs, we save a bit not
having to scan hundreds of srcs when
we're only interested in less than 10.

2) Minor internal changes to the find-supply
function chain, in that rather than passing
a SupplyStore, we actually pass along a map
pointing the the categories of supply.  This
is a precursor to defining categorical supply
relaxations for alternate categories.