Multiple patches to querying and new test
1) Finished hooks for computable supply
categories.  We now have a simple interface
for defining functions that can compute a supply.

a) Integrated lazy map-based merging using
duh - lazymaps - to allow a simple declarative
"joining" of all the supply, without paying the
cost of merging every SRC unless we use it.

b) Removed unnecessary sorting and weight
computations from find-feasible-supply.  We
handle all of that in ces.query/->deployers.

2) Fixed, rather "clarified" incorrect sorting
behavior for logical predicates in ces.query.

a) Clojure's default weight for logical preds
is (< false true).  The context we're using
inline function predicates in (supply ordering
queries) intends the opposite.  Predicates
are inteded to behave more like filter's
behavior, where (< true false).  When we
filter on a compo, for instance, to indicate
preference, we'd like true values to indicate
earlier ordering in the resulting sort.

b) Current solution is to replace calls to
'= and 'not= with 'is and 'is-not, both of
which are simple wrappers that invert the
result of the call.  It's trivial, but helps
the consumer of the api avoid mental gymnastics
due to sorting minutae.  API may change in future
but it works for now.

c) Fixed instances where 2a were causing invalid
sort order, and added a test case.

3) New tests for computed supply categories in
marathon.ces.testing.

a) Querying for nonboggable supply using the existing
infrastructure, ordering according to hld - which
has a component preference - and checking the
composition and order of the result.  This
covers all of the problems in 2.