Fixed small error in demand sampling
marathon.ces.fill.demand/fill-category was
returning only :filled/:unfilled for the
result of filling a demand.  This led to
no recording of a change in the demand
if we partially filled, leading to missing
demandtrend samples reflecting erroneous
missed demands.  Demands looked unfilled
when in fact the fill had changed, albeit
not to completely fill the demand.

Now, we return :filled/:changed/:unfilled,
remaining logic works as advertised,
demands are sampled accordingly.