Minor fixes to ordering criteria compatibility
Minor bug in in source priority resolution was
leading to an absence of applied sorting
criteria, falling back on the default
uniform sort.  This manifested as M3
showing a clear preference for a component
where M4, despite having the same preference
specified in the demand, continued to
uniformly allocate across components.

Problem rested in a typo during conversion
from legacy code.  COMPO_Preference (M3), vs
compo-preference (M4).  Handled this by
injecting a memoized function that parses
the source-first criteria prior to resolving it.
Strings are lower-cased and underscores changed
to hyphens.  Memoization means this only happens
the first time, so little cost involved.

Now, we're seeing the appropriate compo
preferences in M4, the only discrepancy is
a difference in subsequent dwell before deployment
in later NonBOG demand.