Minor tweaks, SRM related
Preparing for a quick visualization
demo, I noticed that the initialization
for units, particularly dwell before
deployment, all started at 1.  Some of
this was due to artificiality in the
input data, but we're supposed to be
computing an accurate cycletime depending
on the position of the policy regardless.

Currently, this problem only really
effects SRM and other highly-specific
supply simulations with initial conditions.

I fixed
marathon.ces.behavior/compute-state-stats
to enforce the original intent, mapping
position in policy to cycletime, as
well as covering the possibility of
non-cyclical initial conditions like
in SRM.  If there exists a path from
the position back to the start state,
whether the path is part of the cycle
or not, we allow the condition to
stand with 0 duration.  Otherwise,
error out.  Now initial conditions
are respected, and show up in the
animated output.