Isolated regression to compute-state-stats
marathon.ces.behavior/compute-state-states
was changed to enable more accurate initial
conditions for SRM.  This caused a regression
in the spawning behavior for M4, that only
showed up in graphical testing (regression
tests missed it).

Pulled in the old "working" function,
currently a/b testing with new function
to make it consistent between M3/M4.