Added finer-grained step functions for vnv
Non-BOG demands, in the presence of other
demands, are causing disparities between
M4 and m3.  M4 is currently illegally allowing
the use of followon supply from other
demand categories, and this is creating
the appearance of preferring AC units over
NG according to the source-first rules.

The crux of the issue appears to be the
current implementation of supply relaxations
for the NonBOG category, which is "too"
relaxed in M4.

To aid in this discovery, I added a couple
of convenience functions that kept popping
up in practice:

1) marathon.ces.engine/parametric-sim-step
Allows callers to define (or omit) the domain
specific functions of the simulation that are
hardwired in engine/sim-step.  This lets us
do things like "step until we start filling".

2) marathan.analysis/step-prefill, fill-of
Higher order convenience functions that build
on 1.  These let us quickly get the context
just prior to fill, to analyze the results
of supply queries and the like.