patched marathon.ces.unit/deployed?
deployed? predicate was implemented incorrectly.
Rather than checking the state-set in the case the
unit's state was a set, it was calling itself
recursively expecting a unit arg vs a state.

1) added aux function, deployed-state? to
clean up the predicate.  This is ::state->boolean

2) Replaced marathon.ces.unit/deployed? with new
implementation using 1).

Appears to have fixed the problem of deployment
feasibility for nonbog, thus fixing the problem
of apparent deployment preferences being illegal.

We're getting correct deployments for the most
part, but still mismatches in the case of mixed
component supply with nonbog demand.