Fixed policy-change induced cycle termination
Unit 5 was getting (coincidentally) 5 days of
dwell in M3 because it was not going to reset.
Prior to executing the policy change, it
accrued 5 days of dwell, then changed policies.
The change from MaxUtilization to a rotational
policy left it at ct=0, effectively going to
reset.  In M3, we don't actually execute the
movement state pipeline, instead handling
the position, location, and state changes
inside of a fairly bulky PolicyChange_State
transition function.

In M4, we typically compose behaviors, and
thus leverage primitives into complex
behaviors like moving-beh.  Inside moving-beh,
we perform a typical check to see if we are
engaging in a transition that will cause
the unit to finish its current cycle (i.e.
going to Reset or the current policy's
designated start state).

This left us with M4 detecting that, going
to reset, Unit 5 was ending its cycle, thus
starting a new cycle (prior to deployment),
effectively erasing the 5 days of accrued
dwell.  Thus, we ended up with 5 days less
dwell before deployment in M4.

The patch adds an additional predicate to
the finish-cycle behavior, which will
bail out if we have a :policy-change
in the environment (that is, we are
downstream of an apply-policy-change
behavior invocation).