Fixing new nonbog selection, vnv by branches now
Added vnv functions for looking at branch-wide
comparisons for vnv.  Base unit types are working
perfectly, but now first of the 01 series is off.
THere appears to be a nonbog selection criteria
difference in M3 vs M4, specifically for pearl on
day 822.  M3 has no available supply, looks like
we aren't deriving an category for finding extra
supply, since there's no SRC in the supplyset (0
supply to start with...)

M4 is actually correct in this aspect, with the
exception of allowing overlapping units to
deploy to non-bog.  Overlapping units are
now considered in the marathon.ces.unit/deployed?
predicate.