Corrected source-first error creating demand ents
1) During testing, certain cases where
multi-compo supply was filling a demand with an
NG_First source-first preference were ending up
with non-NG supply being preferred, like RC.

It turns out, the comparator being using for
the :source-first rule for the demand was
actually the :uniform rule, the default rule
for all demands (in case no source-first rule
is specified).  There was a duplicate argument
in the marathon.data.store/demand declaration,
via spork.ces.entitysystem.store/defentity form
which builds a map-making function under the hood.

The arg 'source-first existed in both the main
arglist and the {:keys [...]} arg list, which
meant that the main arg (actually being passed
in from the entityfactor) would never be used,
instead 'source-first would be bound to nil or
whatever was (never in practice) passed in
via the :source-first binding.

This prompted a patch to spork, bumping the
dependency.  We now detect duplicate argument
cases in defentity, although Clojure allows
this behavior.  I will file a support issue
with clojure.core devs.

2) Also decoupled much of the hairy functions
in marathon.ces.query, particularly for the
ordering functions.  Significantly refactored
them for re-use, and rewrote 'find-supply to
use the simpler functions.  Code clarity
improved, as well as testability.  This led me
to verifying the ordering function ng-first
was actually behaving properly, just never
actually applied...

3) Bumped spork dependency to account for the
patch to defentity.