Handling feaux scientific notation errors
1) Excel likes to be "helpful" by parsing
string values not prepended with ' that
start with E as possible scientific
notation for numbers.  Some srcs are
parsed incorrectly, primarily manifesting
in deployments and demandtrends.  We have
several of these in the test dataset,
so to prevent false-positive diffs,
M3-v80 now intermediately encodes
specific columns with the quotation
in excel, so that the textual output
retains its string encoding.

2) M4 splices in an override for
the default parsing function in incanter
to ensure we don't parse scientific notation.
Otherwise, we end up having no deployments
show up in the dwell plot (another helpful
feature that hinders us in practice...)

3) Implemented a more useful group-based
deployment diff function that is helpful
for looking for outliers in batches of
SRCs, typically applied by branch.