Added recursive guard for policy changes
Rather than using :policy-change key
association existence to determine if
we should try to apply a deferred
policy change, we have a separate
guard :changed-policy.  If
:changed-policy exists, we enforce
the invariant that we only have
one policy change per update cycle.

This addresses an issue introduced
by an earlier revision, which had a
unit changing policies at reset,
according to a deferred policy change,
and actually reverting to its old
policy.

Note: add an invariant in testing
that policy changes can never
cause a unit to revert; or that
units should not change policies
if the change sends them to the
same policy (obvious).