Implemented random weighting for fills
Addressing problems with using unit-index for
deterministic tie-breaking fill criteria.

We now have a field, :unit-weight, which is
derived from a consistent psuedorandom stream
seeded at 0, via the built-in prng.  When
we generate entities in ces.entityfactory,
as a final step, we tack on weights for
the entities derived from this consistent stream.

Had to alter test input, as expected, where two
entities were identical.  The previous index-based
test had the unit with the lower index first.
Now, with the random weighting, the lower-index
unit just happens to have a higher weight, so it's
later now.  Works as expected.   All tests passing.