Patched marathon.ces.unit/cycle-stats for FCLG
As defined in commit bbe385e5, implemented
a minimally invasive patch to alter the
statistics for cycle length, really :max-dwell
, that marathon.ces.unit/normalized-dwell
relies on.  Rather than using the default
max-dwell paramter associated with a policy,
we now ensure that it's not considered
infinite via core/infinite-else, preferring
the value core/+default-cycle-length+,
default is 1095, to provide a finite proxy
for determining normalized dwell for infinite
cyclelengt policies.  This should create a
stable playing field for executing the
comparisong for max-normalized dwell
across heterogenous mixtures of finite and
infinite policies.