Returning wrong convergence when lowbound is non positive
When peak demand is 1, there was no requirement being generated.
This was because given an interval of low and high to search
between, we take the midpoint and define a subinterval from there.
If there is a step length of 1 between low and high [0 1], then
the midpoint calculation returns 0.  In this case, we know we
won't meet anything at 0, so we should converge at 1 instead of at
0.  This patch only affects when low is nonpositive (always 0 for
m4).  If low is nonpositive, we know we converged left since
low is the resulting calculation of a previous high_where_we
missed_demand + 1.