Added simple RA fix to address RA fencepost errors
We ran into a complication when computing or otherwise inferring "misses" from
the original default-distance function. When no supply is present (in the case
of no initial supply and a growth-step of 0), we end up with nil returned
since history->ghosts returns nil if no misses exist, potentially misleading
about the result of the misses.

The fix is to inject some additional information into the requirements
context (that from which we derive a simulation context to test our
growth step).  We add information about the peak demand, under
:peak, and that provides a basis for - later in the pipeline -
calculate-requirement to make a more informed decision about
the results of a requirements run.

If the context to-be-tested (the context resulting from distributing
a sample growth step to the requirements context) contains
some supply, we run through computing misses as before, using
the user-supplied distance function.  We may return nil
as a valid result, indicating no misses.

If no supply exists (possible with a growth-step of 0 and
an empty initial/minimum supply), we now return the peak
demand as the default expectation of missed demand.

This should account for ~4 covergence cases for requirements
analysis, to include 0-growth steps with and without supply.

Also patched erroneous lower to upper in a branch of the
bisecting convergence logic.

Added comments throughout to aid in discovery.