Large update addressing LastDayDefault param
There was a disconnect in the behavior for
the user-defined parameter LastDayDefault
from legacy M3.  The plumbing was in
but not connected properly.  We should
have incorporated the last day parameter
into our determination of the final
simulation day in ces.engine/keep-simulating?
and other places, but did not.

Note: this is different than the tmax
parameter users can provide for
the marathon.analysis/marathon-stream
function.  It's more fundamental to
the generating function of the engine,
although the effect is similar.

Here, marathon.ces.engine/keep-simulating?
consults some internal metrics to determine
if we should continue stepping.

One of them is the behavior for truncating
simulations based on common criteria:
last demand deactivation time,
and user-defined last day.

The LastDayDefault parameter is
now working as intended, and
provides a way to truncate the
simulation (assuming the value
is less than the last demand
deactivation).

During capacity runs, users now
see additional information for
the run, regarding the computed
tfinal (derived from LastDayDefault)
and the empirically-defined
last-deactivation time.  This
is to provide the user with a
visual cue of the implications
of truncation (the min value
of the times will serve as the
cutoff point for truncating the
simulation).

If the user leaves LastDayDefault
blank in parameters, a default
value of 13*365 is assumed.

Tested with requirements analysis
as well.  RA similarly respects the
truncation parameters.