Patched deltaT computations and record lerping
Per user request, we now have the option to
lerp the records rather than computing
purely a deltaT for demandtrends.

The default behavior is bound to
marathon.analysis/*lerp-demandtrends*,
and can be overriden (if pushing out
extra lerped records makes a difference
in time, or memory or whatever).

In practice, not lerping saves
quite a bit of memory and some time
in creating redundant records for
non-sampled days, but it's relatively
neglible (not a bottleneck).

Also, fixed the deltaT computation
to be associated with the actual
record that constitutes a weighted
sample, rather than the record "after."

We can go either way.