beta spec-based validation
We have a working prototype for
data validation based on the clojure
spec library in marathon.specs.

Due to ongoing integration with
the spork.util.parsing and table
libraries, and existing m4 schemas,
we're still not integrated into the
main project workflow.  However,
users can attempt to validate tables
based on m4 schemas (and specs
automatically derived from schemas)
using marathon.specs/validate-records.

Depending on performance, we may
keep this as an error/debug
function meant to be run infrequently
as a sanity check/project validation
tool.  Still, it provides valuable
feedback in where and why input data
is invalid.  Performance appears to be
fast enough for per-project validation
on load without inconveniencing the
user.