Consistency patch for mislabeled fields
Found some erroneous (if unimpactful)
field labels that had worked into the
source and were being overlooked by
the schemas, until running a new
version of proc and running new
validation routines using
clojure.spec.

Typical culprits are:

Spawntime => SpawnTime
Original? => Original

We're getting some inflexibility
in spec at the moment, to include
expectations about strings
and the like for many fields.

If we define a field as text
in the schema, it gets left alone
parsed by identity.  If it's
not actually text though,
the spec thinks it's supposed to
be parsing a resulting string value.

What we really want to say is "any"
instead of the default :text
field.  Temporarily removed problem
fields from the specs to get things
working.

One of the cool things is that
when input validation fails,
we get a thorough report on
where (row) and why (field(s)
and spec explanation).