Fixed demandtrends sampling, deltaT, and more.
Patched a problem when computing deltaT from
adjacent demandtrends samples.  We were naively
using adjacent deltaT without accounting for
finalized, or day-before-deactivation demand
trends.  These samples are implicitly meant to
carry a weight of 1, i.e. deltaT = 1.

We were blowing them out of proportion
by overweighting the deltaT, and ignoring
the finality of said demand, thus making it
appear that the demand was continuing well
past its deactivation date.

We now explicitly check for finalized demands,
and when creating the demandtrend, provide
an pre-computed deltaT for them.  This
signals further down the pipe that the
trend is associated with a finalized
demand, that is, the demand should not
be subject to the deltaT of the larger
inter-group time delta, and that it will
only last 1 day.  Additionally, when
lerping to fill in time intervals,
we only include the finalized
records for the first interval, then
lerp all other demandtrends as normal.

Also fixed a problem where the last
day of demand trends where being ignored.
We now should be accurately sampling
even in truncated simulations.