Pending implementation of finite cyclelength
It looks like we can just wire in some logic
in marathon.ces.unit/cycle-stats, which
funnels the information for
+default-cycle-length+ into the cycle record,
which is used for computing proportions and
the like, and we should be golden. That is, we
can let the default machinery for cycle
transfers leverage the infinite cyclelength
that emerges from the policy naturally (e.g.
there's a state transition somewhere that's
>= 9999999), while using the user-defined
FCLG for normalized dwell purposes to enable
fair scoring. So, this ends up being a
fairly minor additive modification....

Currently storing :finite-cyclelength in
the policy record, but this breaks protocol
idioms...Looking at defining a minimal protocol
that will extend to policies, which will delegate
to looking up this key if it exists, otherwise
fallback to +default-cycle-length+ for legacy
behavior.  Then switch out invocations in
ces.entitybuilder for cycletime distribution,
and ces.unit for cycle-stats.

That should be it for integration.