Candidate patch to fix followons and RC filtering
Problem:
We ran into repeat followon deployments to
nonbog demands causing errors with entities
trying to compute seemingly unnecessary moves.

The root is that, during followon, we put the
entity in a temporary holding state so it
can withdraw from its deactivating demand,
without re-entering or resetting completely.

This supply is then used preferentially to
fill demands with similar demandgroups.

When a unit does successive fills of this
type, the :positionpolicy is effectively
clobbered in the process.  Normally,
in the case where the unit is not used
for a followon (99% of the time),
it goes through recovery or reset and computes
a move explicitly.

In the edge case, the unit will try to
update prior to computing a move, and fail
to do so because it doesn't know where to
go from :followon.  In reality, the
positionpolicy for the unit is whatever
it came into the wait-based move (nonbog)
with....Yet we leverage the policyposition
explicitly to communicate moves and changes
in the behavior system.

Solution:
The best case is to inject a weak recovery into the
marathon.ces.behavior/wait-based-beh, so that - IFF -
we are coming into a wait-based move, e.g. a deployment
to a nonbog event, from a :followon status, derived
via :policyposition, then we compute the actual
position we're supposed to be waiting in, and schedule
a position change (from :followon->wait position)
during the update.

This does a little extra work, but shouldn't
be called too often.  It also only covers the
corner case for followon deployments to
nonbog, or wait-based movements, where
we want to retain the expected policy position
for units that are paused.  Other units
go on their way to recovery.

This should be broadly compatible in both
directions (e.g. nonbog -> bog, in the case
where a unit was drawn from available
forces by happenstance, that unit should
be able to interchangeably followon to
other bog demands (e.g. demands with
no rules mapping to effects that convey
an infinite wait time and a :waiting
state).