Added hooks for expressing modernization
Made multiple patches to input processing,
source first criteria, demand categories,
and entity behavior to express
the notion of modernization.

Modernization can currently be expressed
as a sourcing criteria, and as a category
of demand.

As a demand, a "mod demand" acts similar
to a nonbog demand, in that units go
through a pseudo-deployment where
they really just wait out-of-cycle
at some location.  While at this location
their state is #{:waiting :modernizing}.

The :modernizing state has a new behavior
that, upon completion of the unit's
duration, will improve the mod-level
of the unit (e.g. 2->1).