Patching in better mechanism for demand filters
We have a corner case with the non-bog demands
that we're using for modernization rules.

Trying to implement AC-only modernization,
which works great if you plug in compo
preference for the computed supply.

This still admits non-AC units from the
:default category, which is unfiltered.

We can trivially attack this via a top-level
filter in marathon.ces.fill/demand->rule
that associates with the :where clause of
the demand rule.

Perhaps even better is to just inline
the damn rule, e.g.

{:rule Modernization :Where [AC]}

This would be simple enough...