Added demand filters, proto query dsl
We were missing the mapping of
demand Categories to actual filters.

The underlying query infrastructure
had this in place, and we
actually use it (in addition to
computed supply from NonBOG categories),
but we missed the linkage to general
demand categories.

We now have a (temporary) map in
marathon.ces.fill/demand-filters,
which informs marathon.ces.fill/demand->rule,
which allows us to define demand filters
and apply them automatically.

Currently, the only active filter is for
Modernize-AC.