Refactored marathon.analysis.random/fill-stats, +
Broke out the functions in fill-stats, removed
the old call to remove cannibalization demands
from the input.  We now determine accumulation
behavior based on DemandGroup, and if it's
registered in the
marathon.analysis.random/demand-group-type map.

If it's not, we accumulate fill as before,
adding stats and demand quantity to the fill-stats
accumulator.

If it is, we accumulate the fill stats by compo
into not-ready by compo in the accumulator,
and do not accumulate total-quantity for
the demand.  This will respect the intent
of certain demand types (e.g. RC_NonBOG-War)
that absorb units without counting toward fill,
by accounting for them as not-ready and by not
increasing the observed demand.

This should fix the accounting errors we saw
with certain RC units that had these kind of demands,
and fit with the sponsor's intent for the fill stats.